painter/requirements.txt
PyQt5==5.9.2
fbs==0.8.6
numpy==1.15.4
setuptools==46.1.3
natsort==6.0.0
scikit-image==0.14.0
PyWavelets==1.1.1
Pillow==6.0.0
scipy==1.3.0
